+++ server/source_layout_test.go
@@ -16,7 +16,7 @@ func TestServerRoutesAreSplitByGraphPort(t *testing.T) {
 		{"routes_write.go", []string{"func (s *Server) handleSemanticMutation", "func (s *Server) handleCreateStructure"}},
 		{"routes_unixfs_compat.go", []string{"func (s *Server) handleWrite", "UnixFSWriteResponse"}},
 		{"routes_resolve.go", []string{"func (s *Server) handleResolve", "func (s *Server) serveResolve"}},
-		{"routes_verify.go", []string{"func (s *Server) handleVerify", "authverifier.NewDefault", "portable.VerifyProofList"}},
+		{"routes_verify.go", []string{"func (s *Server) handleVerify", "s.verifierCache.load", "portable.VerifyProofList"}},
 		{"routes_content.go", []string{"func (s *Server) handleContent", "func (s *Server) readContentPayload"}},
 		{"routes_admin.go", []string{"func (s *Server) handleHealth", "func (s *Server) handleMetrics"}},
 	}